# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NextPressKit API
  version: "0.1.0-phase4"
  description: |
    Authentication supports two modes (configure with `JWT_AUTH_SOURCE` on the server):

    - **cookie** (default): access and refresh tokens are HttpOnly cookies (`JWT_ACCESS_COOKIE_NAME`, `JWT_REFRESH_COOKIE_NAME`, default `access_token` / `refresh_token`). Login and refresh responses return `user` only; browsers send cookies automatically when `CORS_ORIGINS` is set and the client uses credentials.
    - **header**: tokens are returned in JSON on login/refresh; send `Authorization: Bearer <access_jwt>` on protected routes.

    Protected routes accept **either** Bearer JWT **or** the access-token cookie (when the server is in cookie mode and the client sends cookies).
servers:
  - url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: Access JWT in HttpOnly cookie (default cookie name; override with `JWT_ACCESS_COOKIE_NAME`).
  schemas:
    errorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    authUserPublic:
      type: object
      properties:
        id: { type: integer, format: int64 }
        uuid: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        active: { type: boolean }
        rbac:
          type: object
          properties:
            roles:
              type: array
              items: { type: string }
            permissions:
              type: array
              items: { type: string }
          required: [roles, permissions]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
      required: [id, uuid, firstName, lastName, email, active, rbac]
    authTokensWithUser:
      type: object
      description: |
        When `JWT_AUTH_SOURCE=cookie` (default), login/refresh responses omit `tokens` and set cookies instead.
        When `JWT_AUTH_SOURCE=header`, `tokens` is present with access and refresh JWT strings.
      properties:
        tokens:
          type: object
          nullable: true
          properties:
            accessToken: { type: string }
            refreshToken: { type: string }
        user:
          $ref: "#/components/schemas/authUserPublic"
      required: [user]
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: { type: string }
                last_name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
              required: [first_name, last_name, email, password]
      responses:
        "201":
          description: user registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer, format: int64 }
                  uuid: { type: string }
                  firstName: { type: string }
                  lastName: { type: string }
                  email: { type: string }
                required: [id, uuid, firstName, lastName, email]
        "400":
          description: invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /auth/login:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        "200":
          description: |
            Current user profile. In cookie mode, also sets HttpOnly `access_token` and `refresh_token` cookies.
            In header mode, includes `tokens.accessToken` and `tokens.refreshToken` in the JSON body.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Present when `JWT_AUTH_SOURCE=cookie` (may appear twice for access and refresh cookies).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/authTokensWithUser"
        "401":
          description: invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /auth/me:
    get:
      tags: [auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: Current user from access token
      responses:
        "200":
          description: authenticated user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/authUserPublic"
                required: [user]
        "401":
          description: missing or invalid access token (Bearer or access cookie, depending on server config)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: user id in token no longer exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /auth/refresh:
    post:
      tags: [auth]
      requestBody:
        required: false
        description: Required when `JWT_AUTH_SOURCE=header`. Omit body in cookie mode (refresh token is read from the `refresh_token` cookie).
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: |
            Current user profile. In cookie mode, rotates HttpOnly cookies. In header mode, includes new `tokens` in JSON.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Present when `JWT_AUTH_SOURCE=cookie`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/authTokensWithUser"
        "401":
          description: invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
  /auth/logout:
    post:
      tags: [auth]
      requestBody:
        required: false
        description: Required when `JWT_AUTH_SOURCE=header`. Omit body in cookie mode.
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Logout accepted; in cookie mode clears auth cookies.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Present when `JWT_AUTH_SOURCE=cookie` (clears cookies).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                required: [message]
        "401":
          description: invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
...